requirements.txt
asgiref
Django
djangorestframework
drf-yasg
inflection
packaging
permission
Pillow
psycopg2-binary
python-decouple
pytz
PyYAML
sqlparse
typing_extensions
uritemplate
djangorestframework-simplejwt
twilio
python-telegram-bot